+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - ServiceProviderを参照する形からコピーする形に変更(以前のServiceProviderは0.7.0で削除される予定)
 - .env.localを生成するようにした
 - config/tucle.phpにcopyrightとpowered_byを追加
+- Bugsnagを追加
 
 ##### アップグレード
 
@@ -32,6 +33,15 @@ $ php artisan tucle:builder module_name --force
   'copyright' => 'Eyewill',
 
   'powered_by' => 'Tucle5.2',
+~~~
+
+- composerとapp.phpを更新
+~~~
+$ php artisan tucle:init --only=composer,config --force
+~~~
+- .envにBUGSNAG_API_KEYをセット
+~~~
+BUGSNAG_API_KEY=*****
 ~~~
 
 #### 0.5.0